rename read-file to read-h5 to make it explicit (#98)
* rename read-file to read-h5 to make it explicit

* fix travis test

* remove append of coverage

* add all the __init__.py files for tests to fix coverage report in travis!

* add back all the append coverage
