+++ requirements.txt
@@ -9,3 +9,4 @@ tifffile
 neuroglancer
 tinybrain
 zmesh
+tifffile
